Use germinate properly. (#6).
We don't want a script to generate the extra files, it's not worth it or 
really used and we shouldn't try it.

Two architectures were removed as in previous commit.

Per usual update.cfg was updated along with debian/control and 
everything was cleaned up.

In hirsute/vcs seed_base will have to stay like this because eventually 
the repos will change to in launchpad/bzr.

metapackage-map was also updated as in part of the issue #6 solving 
commits.